Major updates
The merge and apply_visibility methods have been completely rewritten and simplified. They are now both based on the ticks-method that was used to compute the mutlitple visibility windows in visibility.py (i.e. all interavll edges are ordered in a list and each interval produced in that way are assigned a True or False value depending on some criteria). The plot mehod has been modified to handle multiple windows. The single_site method that was calling apply_visibility has been removed and apply_visibility (new version) is directly called. The both_sites method relies better on apply_visibility.