Improved visibility class
The Visibility is now a standalone class. Two instances, one for North, the other one for South, are attached to the GammaRayBurst class.
All information, in particular the default values, initially in the GRB class have been moved in the Visibility class.
The visibility also handles now the moon veto period.
The computation of the new visibility has been improved and refurbished (including moon periods); Some unused function are ready to be exploited when the Moon veto will be indeed used (so far only the Moon min altitude, that defines the initial moon periods, is used).
The main function has been cleaned-up. It produces updated visibilities for a list of GRB and is able to compare with the former default visibility.