Update configuration.py
Add an option to have the configuration file copied or not (copy is wishable when simulating not analysing)
Define default configuration file name as a global variable. def_file argument is now conf_file (in the siprit of passing any kind of filename)
Reorganise a bit the consequence of saving the simulation (e.g. niter=1)