Update visibility.py
Correct a bug that had no importance as long as the prompt component was not simulated: the vis_prompt flag was not correctly computed.
The documentation and comments on the constructor are improved.
Re-arrange completely the main function to take into account the new configuration scheme (reading from a yaml file instead of getting the parameters from a python module)