Major update to compute the visibility within the GRB (required by the prompt)
The visibility is now obtained from within the .from_fits method in all cases. There was initially the default visibility read from the GRB fits file but this was incompatible with the decision to read the prompt data if the vis_prompt variable is True.
As a consequence, the GRB constructor .from_fits takes now a visibility variable that can have three from : None, the default visibility is read, a keyword among those available in visibility.yaml, the visibility is computed on the fly, a folder name,the visibility is read from this folder. The main function has be reorganized accordingly.