Implement short GRB reading
The short GRB come with a slightly different file format, requiring some corrections. The trigger date is also given as a relative data in JD, thus requiring to fix the t0.
Changes to read either fits or fits.gz file depending on what is available on disk.
Various tests done on the presence of certain keyword in the fits file (short GRB format change).
Flux unit directly read from the fits file.
Logic implemented to get the prompt only if it has a chance to be visible (using vis[].vis_prompt)