Change spectral models into full SkyModel models
Make explicit the reading of the default fits file visibility ("None" was doing the job, now reserved to no visibility. "built-in" is the new keyword).
Now store a list of SkyModel with spectral and spatial models instead of a list of spectral models only. This is useful to dump the models into yaml file and make the code conceptually simpler.
New function model_to_yaml to dump the model into yaml files.